requirements.txt
fastapi
uvicorn[standard]
torch
Pillow
jinja2
python-multipart
torchvision
torchinfo
matplotlib
tqdm
typing_extensions